Improve hasFilters() method on SelectFilter.
Previously, the hasFilters() method on SelectFilter looked for active
filters using the query string.  This failed when a filter had been
added by setting DefaultVars for a field and was the sole need
connecting SelectFilter to the Request.  The new version reuses the
getCurrentFilters() method, which is more reliable.  The older, hacky
version is still present and is used if the caller doesn't pass a Fields
object to hasFilters().  This should take care of any backwards
compatibility issues.